requirements.txt
MouseInfo==0.1.3
numpy==2.2.5
opencv-python==4.11.0.86
pillow==11.2.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==11.0
pyobjc-framework-ApplicationServices==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-CoreText==11.0
pyobjc-framework-Quartz==11.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
rubicon-objc==0.5.0
setuptools==80.1.0
six==1.17.0
wheel==0.45.1
